requirements.txt
redis==5.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
numpy>=1.26.0